requirements.txt
ShopifyAPI==12.0.0